+++ lib/src/core/logger.dart
@@ -1,3 +1,5 @@
+import 'package:ai_localizer/ai_localizer.dart';
+
 /// Simple logger with support for different log levels
 class Logger {
   final bool verbose;
@@ -6,28 +8,28 @@ class Logger {
 
   /// Log an informational message
   void info(String message) {
-    print('\x1B[34m[INFO]\x1B[0m $message');
+    logger.info('\x1B[34m[INFO]\x1B[0m $message');
   }
 
   /// Log a debug message (only shown in verbose mode)
   void debug(String message) {
     if (verbose) {
-      print('\x1B[36m[DEBUG]\x1B[0m $message');
+      logger.info('\x1B[36m[DEBUG]\x1B[0m $message');
     }
   }
 
   /// Log an error message
   void error(String message) {
-    print('\x1B[31m[ERROR]\x1B[0m $message');
+    logger.info('\x1B[31m[ERROR]\x1B[0m $message');
   }
 
   /// Log a warning message
   void warning(String message) {
-    print('\x1B[33m[WARNING]\x1B[0m $message');
+    logger.info('\x1B[33m[WARNING]\x1B[0m $message');
   }
 
   /// Log a success message
   void success(String message) {
-    print('\x1B[32m[SUCCESS]\x1B[0m $message');
+    logger.info('\x1B[32m[SUCCESS]\x1B[0m $message');
   }
 }